Implemented rawmemchr2
This is a routine similar to glibc's rawmemchr(3), except that
it scans for two characters, c1 and c2, whichever comes first.
The search is unbounded, which means that the caller should
typically install a sentinel (c1 or c2) at the end of the buffer.

Furthermore, when possible, the routine exercises unaligned reads
of 16/8/4 characters at a time; thus the last read can cross a page
boundary, which can trigger a segmentation violation.  This means
that, after the sentinel, the caller must also provide 15 bytes
of padding.