Optimize cpp output parsing, use rawmemchr2
cpp output can be huge, and parsing it one character at a time,
in a loop which is not exactly tight, is tantamount to another
preprocessor run (the real preprocessor is still much slower,
though, and needs to be optimized as well.)

$ cat gtkmm.cc
 #include <gtkmm.h>
$ g++-6 `pkg-config --cflags gtkmm-3.0` -E gtkmm.cc |wc -c
6876510

It seems that parsing can only be improved by skipping a few
bytes at a time.  The rawmemchr2 routine has been specifically
crafted for this task.  Working as a loop control, it skips towards
the next "#" or "." character (the latter to check for the .incbin
directive).  (This makes the code somewhat less amenable to certain
future possibilities.  What if another character needs to be treated
specially?  Well, a rawmemchr3 routine can be crafted in a rather
straightforward manner.)

Another important optimization is to avoid copying file paths and
calling remember_include_file() all over again.  This might seem like
a minor issue at first, which it isn't: there's a hell of a lot of
those paths, and the costs do add up.

$ g++-6 `pkg-config --cflags gtkmm-3.0` -E gtkmm.cc |grep ^# |wc -l
20083
$ g++-6 `pkg-config --cflags gtkmm-3.0` -E gtkmm.cc |grep ^# |
	cut -d' ' -f3 |sort -u |wc -l
1427

This shows that malloc/strcpy/murmurhash is getting called 14 times
more than strictly necessary.  Most of those calls can be avoided
by cooperating with the preprocessor: either expecting a special flag,
or simply checking if the line number is 1.  (This optimization relies
on a preprocessor being sane.  Well, this is already essential for
ccache to work anyway.)

There are also a few minor optimizations, each producing a statistically
measurable difference in terms of both valgrind instruction reads and
rdtsc counter cycles.

As a result, ccache now runs 2+ times faster, as per valgrind
instruction reads.

$ CCACHE_NODIRECT=1 valgrind --tool=callgrind \
./prev/ccache g++-6 `pkg-config --cflags gtkmm-3.0` -c gtkmm.cc
==900== I   refs:      134,707,769
$ CCACHE_NODIRECT=1 valgrind --tool=callgrind \
     ./ccache g++-6 `pkg-config --cflags gtkmm-3.0` -c gtkmm.cc
==907== I   refs:      61,176,244

The improvement in the actual running time is also noticable
if dwarfed by the running time of cpp.  Here are some typical
numbers.

$ time pkg-config --cflags gtkmm-3.0 >/dev/null
0.12s user 0.00s system 99% cpu 0.126 total
$ cflags=`pkg-config --cflags gtkmm-3.0`
$ time g++-6 $cflags -E gtkmm.cc >/dev/null
0.32s user 0.04s system 99% cpu 0.356 total
$ time CCACHE_NODIRECT=1 ./prev/ccache g++-6 $cflags -c gtkmm.cc
0.34s user 0.06s system 99% cpu 0.401 total
$ time CCACHE_NODIRECT=1      ./ccache g++-6 $cflags -c gtkmm.cc
0.33s user 0.05s system 99% cpu 0.376 total

According to somewhat more precise measurements, the improvement
the in running time is about 5%, or about 0.02s per run.  Note that
roughly half of the ccache overhead has been shaved off; the rest
is now dominated by MD4, comprising as much as 76%.

There are also some minor changes to the parser logic; noticeably,
what constitutes an error.  Previously, finding a newline before
the opening quote was not an error, but finding EOF was.  The latter
case is now treated the same way as the former.  The real error is
when the opening quote has been found, but then a newline (or EOF)
is encountered before the closing quote.  (Before this change,
the parser could actually accept the closing quote on a next line.)

There's also a minor restructuring of the code wrt remember_include_file
call.  The checks which are specific to cpp output, such as ignoring
"<command-line>" filenames and taking into account system header
sloppiness, have been taken out and are now performed in the parsing
code, mostly to avoid the call.  However, these checks actually
do not to apply to filenames which come from the command line,
such as included_pch_file which is used as an argument to another
remember_include_file call.

Well, actually I need this change for something rather different:
I want to process more cpp directives.  Thus q, the input pointer,
first must be positioned at a "#" character which is also at the
beginning of a line.  Directives then can be processed in a row.
But I couldn't help but introduce more optimizations.